Renamed litexcnc_spi.* tp litexcnc_spidev.*
